+++ changelog.md
@@ -1,3 +1,9 @@
+# v1.1.0 
+- Bumped Geode loader support to `v5.8.0`
+- Added 10 new notifications
+- Added 15 new characters
+- Fixed issue with notifications continuing after level completion
+
 # v1.0.1 
 - Fixed issue with responding to update forecast notification
 - oops forgot credits sry